+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to insta and cargo-insta are documented here.
 ## 0.18.0
 
 - `Settings::bind` now can return a result.
+- Expose the drop guard type of `bind_to_scope`.
 
 ## 0.17.1